+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to kryten-cli will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.4] - 2026-03-14
+
+### Added
+
+- **About Command**: Added `kryten about` to display project metadata
+  - Project name and current version
+  - Credits and project link
+  - Thank-you message for the people who make Cytu.be's Channel-Z excellent
+
 ## [2.6.3] - 2026-03-14
 
 ### Fixed